refactor(cli): split main.rs into commands/ module directory
Modularize the 2222-line main.rs into focused submodules for
maintainability. No logic, behavior, or API changes -- pure
restructuring with identical test results (71 tests pass).

New structure:
  main.rs             -- CLI definitions, dispatch, daemon, shared helpers
  commands/
    mod.rs            -- submodule declarations and re-exports
    parse.rs          -- parse, condition, stdin subcommands
    validate.rs       -- validate subcommand
    lint.rs           -- lint subcommand, schema validation, Painter
    eval.rs           -- eval subcommand, stream/line evaluation helpers
    convert.rs        -- convert, list-targets, list-formats subcommands
