+++ crates/rsigma-cli/src/commands/mod.rs
@@ -1,0 +1,11 @@
+mod convert;
+mod eval;
+mod lint;
+mod parse;
+mod validate;
+
+pub(crate) use convert::{cmd_convert, cmd_list_formats, cmd_list_targets};
+pub(crate) use eval::cmd_eval;
+pub(crate) use lint::cmd_lint;
+pub(crate) use parse::{cmd_condition, cmd_parse, cmd_stdin};
+pub(crate) use validate::cmd_validate;
